Rename `handlers` module into `tasks`
